requirements.txt
fastapi
uvicorn
python-multipart
reportlab
python-dotenv
marker-pdf
openai
torch